hot fix

+++ src/main/java/com/bookingtrips/booking_trips_backend/service/FavoriteService.java
@@ -3,6 +3,8 @@ package com.bookingtrips.booking_trips_backend.service;
 import com.bookingtrips.booking_trips_backend.dao.FavoriteRepository;
 import com.bookingtrips.booking_trips_backend.dto.TripDto;
 import com.bookingtrips.booking_trips_backend.entity.Favorite;
+import com.bookingtrips.booking_trips_backend.exception.FavoriteAlreadyExistsException;
+import com.bookingtrips.booking_trips_backend.exception.ResourceNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +15,7 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 public class FavoriteService {
+
     private final FavoriteRepository favoriteRepository;
     private final AuthenticationService authenticationService;
     private final TripService tripService;
@@ -20,18 +23,23 @@ public class FavoriteService {
     @Transactional
     public Favorite addFavorite(Long tripId) {
         Long userId = authenticationService.getLoggedInUser().getUserId();
+
         if (favoriteRepository.existsByUserIdAndTripId(userId, tripId)) {
-            throw new IllegalStateException("Trip is already in favorites");
+            throw new FavoriteAlreadyExistsException("Trip is already in favorites");
         }
+
         Favorite favorite = new Favorite();
         favorite.setUserId(userId);
         favorite.setTripId(tripId);
+
         return favoriteRepository.save(favorite);
     }
 
     public List<TripDto> getUserFavorites() {
         Long userId = authenticationService.getLoggedInUser().getUserId();
+
         List<Favorite> favorites = favoriteRepository.findByUserId(userId);
+
         return favorites.stream()
                 .map(favorite -> tripService.getById(favorite.getTripId()))
                 .collect(Collectors.toList());
@@ -40,6 +48,12 @@ public class FavoriteService {
     @Transactional
     public void removeFavorite(Long tripId) {
         Long userId = authenticationService.getLoggedInUser().getUserId();
+
+        boolean exists = favoriteRepository.existsByUserIdAndTripId(userId, tripId);
+        if (!exists) {
+            throw new ResourceNotFoundException("Favorite not found for the given trip");
+        }
+
         favoriteRepository.deleteByUserIdAndTripId(userId, tripId);
     }
 }
